Struct oniguruma::Regex

pub struct Regex {
    // some fields omitted
}

A compiled Oniguruma regular expression.

Methods

impl Regex

fn new(pattern: &str) -> Result<Regex, Error>

Compiles a regular expression with default options. Default syntax is SYNTAX_RUBY.

Once compiled, it can be used repeatedly to search in a string. If an invalid expression is given, then an error is returned.

fn new_with_option(pattern: &str, option: Options) -> Result<Regex, Error>

fn new_with_syntax(pattern: &str, syntax: *const OnigSyntaxTypeStruct) -> Result<Regex, Error>

fn new_with_option_and_syntax(pattern: &str, option: Options, syntax: *const OnigSyntaxTypeStruct) -> Result<Regex, Error>

fn search(&self, text: &str, region: &mut Region, option: Options) -> Result<Option<usize>, Error>

Search pattern in string and store search result into region object.

You also can use search time options: OPTION_NOTBOL and OPTION_NOTEOL.

fn captures<'t>(&self, text: &'t str) -> Option<Captures<'t>>

Returns the capture groups corresponding to the leftmost-first match in text. Capture group 0 always corresponds to the entire match. If no match is found, then None is returned.

Panics

This method may panic in the case memory overflow during execution or other internal errors of Oniguruma engine.