Crate oniguruma

Rust bindings for the Oniguruma regular expressions library.

Example of usage:

use oniguruma::Regex;

let regex = Regex::new("e(l+)").unwrap();
for (i, pos) in regex.captures("hello").unwrap().iter_pos().enumerate() {
    match pos {
         Some((beg, end)) =>
             println!("Group {} captured in position {}:{}", i, beg, end),
         None =>
             println!("Group {} is not captured", i)
    }
}

Structs

Captures
Captures represents a group of captured strings for a single match.

Error
An error that occurred during parsing, compiling or evaluating a regular expression.

Options
Regex parsing, compilation and evaluation options.

Regex
A compiled Oniguruma regular expression.

Region
Representation of regex search result.

SubCaptures
An iterator over capture groups for a particular match of a regular expression.

SubCapturesPos
An iterator over capture group positions for a particular match of a regular expression.

Constants

OPTION_CAPTURE_GROUP
Named and no-named group captured. This is search time option.

OPTION_DONT_CAPTURE_GROUP
Only named group captured. This is search time option.

OPTION_EXTEND
Extended pattern form. This is compile time option.

OPTION_FIND_LONGEST
Find longest match. This is compile time option.

OPTION_FIND_NOT_EMPTY
Ignore empty match. This is compile time option.

OPTION_IGNORECASE
Ambiguity match on. This is compile time option.

OPTION_MULTILINE
'.' match with newline. This is compile time option.

OPTION_NEGATE_SINGLELINE
Clear OPTION_SINGLELINE which is enabled on SYNTAX_POSIX_BASIC, SYNTAX_POSIX_EXTENDED, SYNTAX_PERL, SYNTAX_PERL_NG, SYNTAX_JAVA. This is compile time option.

OPTION_NONE
Default options. This is both compile and search time option.

OPTION_NOTBOL
String head isn't considered as begin of line

OPTION_NOTEOL
String end isn't considered as end of line

OPTION_SINGLELINE
'^' -> '\A', '$' -> '\Z'. This is compile time option.

Statics

ENCODING_UTF8

SYNTAX_ASIS
Plain text syntax

SYNTAX_EMACS
Emacs syntax

SYNTAX_GNU_REGEX
GNU regex syntax

SYNTAX_GREP
Grep syntax

SYNTAX_JAVA
Java (Sun java.util.regex) syntax

SYNTAX_PERL
Perl syntax

SYNTAX_PERL_NG
Perl + named group syntax

SYNTAX_POSIX_BASIC
POSIX Basic RE syntax

SYNTAX_POSIX_EXTENDED
POSIX Extended RE syntax

SYNTAX_RUBY
Ruby (default) syntax